refactor: restructure AppLayout components to improve organization and maintainability; introduce AppLayoutFooterContext for dynamic footer management; update styles and responsive behavior in SidebarLayout and HeaderLayout
