--- src/components/AppLayout/context/index.ts
@@ -1,1 +1,0 @@
-export { AppLayoutFooterProvider, useAppLayoutFooter } from './AppLayoutFooterContext';
