+++ src/store/appLayout.store.ts
@@ -25,9 +25,14 @@ export interface AppLayoutStore {
 
 	userInformation?: IUserInformation;
 	setUserInformation: (userInformation: IUserInformation) => void;
+
+	collapsed: boolean;
+	setCollapsed: (collapsed: boolean) => void;
 }
 
 export const useAppLayoutStore = create<AppLayoutStore>(set => ({
+	collapsed: false,
+	setCollapsed: (collapsed: boolean) => set({ collapsed }),
 	agencies: [],
 	setAgencies: (agencies: IAgency[]) => {
 		set({ agencies });
